+++ main.py
@@ -62,8 +62,8 @@ class Error_Types(Enum):
     WARNING_FIELD_WITHOUT_AUTHORITY_ID = 30
     RECORD_WAS_NOT_CHANGED = 31
     WARNING_MULTIPLE_AUTHORITY_ID_IN_ONE_FIELD = 32
+    AUTH_ID_HAS_NO_CURRENT_FIELD = 33
     
-
 # ----------------- Classes definition -----------------
 class Error_File(object):
     def __init__(self, file_path:str) -> None:
@@ -118,12 +118,83 @@ class Report_Updated_Bibnb_File(object):
     def close(self):
         self.file.close()
 
+class Preferred_Field(object):
+    """Must be used after ensuring the field has at least 1 $9"""
+    def __init__(self, field:pymarc.field.Field):
+        self.id:str = get_auth_id(field)
+        self.old_field:pymarc.field.Field = None
+        self.current_field:pymarc.field.Field = None
+        self.update_with_new_field(field)
+    
+    def update_with_new_field(self, field:pymarc.field.Field) -> bool:
+        """Updates the authority with new field.
+        Returns if the new field is used instead of the old one"""
+        # If no current field was used, adds it and end here
+        if self.current_field == None:
+            self.current_field = field
+            return True
+        # If the field already existed, checks if there are PPN
+        if len(field.get_subfields("3")) > 0:
+            # Checks if current field has PPN
+            # If not, replace the olf field
+            if not self.has_ppn:
+                self.__replace_current_field(field)
+                return True
+            # If there are PPN, checks which field is the closest to
+            # the same of Koha ID & PPN
+            if not self.nb_ppn_match_nb_ids:
+                # If new field is closer to the perfect match, replace it
+                if abs(len(field.get_subfields("9")) - len(field.get_subfields("3"))) < abs(self.nb_koha_id - self.nb_ppn):
+                    self.__replace_current_field(field)
+                    return True       
+        
+        # PPN have priotity so after it
+        # Keep $7='ba0yba0y' > $7='ba' > other / none
+
+        # By default, return False
+        return False
+
+    @property
+    def nb_koha_id(self) -> int:
+        """Returns the number of $9 in current field.
+        Returns -1 if no current field is defined"""
+        if self.current_field == None:
+            return -1
+        return len(self.current_field.get_subfields("9"))
+
+    @property
+    def nb_ppn(self) -> int:
+        """Returns the number of $3 in current field.
+        Returns -1 if no current field is defined"""
+        if self.current_field == None:
+            return -1
+        return len(self.current_field.get_subfields("3"))
+
+    @property
+    def has_ppn(self) -> bool:
+        """Returns if current field has PPN"""
+        return self.nb_ppn > 0
+    
+    @property
+    def nb_ppn_match_nb_ids(self) -> bool:
+        """Returns if the number of PPN matches the number of IDs"""
+        return self.nb_ppn == self.nb_koha_id
+
+    def __replace_current_field(self, field:pymarc.field.Field):
+        self.old_field = self.current_field
+        self.current_field = field
+
+
 # ----------------- Functions definition -----------------
+def get_auth_id(field:pymarc.field.Field) -> str:
+    """Returns the auth id of a field"""
+    return "-".join(field.get_subfields("9"))
+
 def dedupe_field(record:pymarc.record.Record, tag:str, index:int=None, bibnb:int=None) -> bool:
     """Removes multiple occurence of fields sharing the same $9
     
     Returns a bool to know if the record was edited"""
-    auth_id_index:List[str] = []
+    auth_id_index:Dict[str, Preferred_Field] = {}
     fields:List[pymarc.field.Field] = []
     for field in record.get_fields(tag):
         # If no authority ID, keep the field but log a warning
@@ -134,21 +205,36 @@ def dedupe_field(record:pymarc.record.Record, tag:str, index:int=None, bibnb:int
             continue
         # For info purpose, checks if multiple $9
         # RAMEAU terms might have multiple $9 ($a-$x), with different combination possible
+        # 1.1 : Keeping this behaviour evenn if new class might have tools to deals wiht it better
         if len(field.get_subfields("9")) > 1:
             ERRORS_FILE.write(Error_Types.WARNING_MULTIPLE_AUTHORITY_ID_IN_ONE_FIELD, index=index, bibnb=bibnb, msg=marc_utils.field_as_string(field))
             LOG.record_message(Level.WARNING, index, bibnb, msg=f"Field has multiple authority ID : {marc_utils.field_as_string(field)}")
-        # Checks if this ID was already met
-        # If yes, log an info + report & don't add the field to the list
-        auth_id = "-".join(field.get_subfields("9"))
-        if auth_id in auth_id_index:
-            LOG.record_message(Level.INFO, index, bibnb, msg=f"Deduping on authority ID {auth_id} : {marc_utils.field_as_string(field)}")
-            DELETED_FIELD_FILE.write(bibnb, index, tag, auth_id, field)
+        # Get auth ID to check against index
+        auth_id = get_auth_id(field)
+        # If this auth_id does not exist, adds it to the index
+        if not auth_id in auth_id_index:
+            auth_id_index[auth_id] = Preferred_Field(field)
+        # If it does exist, dedupes it
+        else:
+            changed = auth_id_index[auth_id].update_with_new_field(field)
+            deleted_field = field
+            if changed:
+                deleted_field = auth_id_index[auth_id].old_field
+                LOG.record_message(Level.INFO, index, bibnb, msg=f"Replacing preferred field for authority ID {auth_id} from {marc_utils.field_as_string(auth_id_index[auth_id].old_field)} to {marc_utils.field_as_string(auth_id_index[auth_id].current_field)}")
+            # Log an info + report
+            LOG.record_message(Level.INFO, index, bibnb, msg=f"Deduping on authority ID {auth_id} : {marc_utils.field_as_string(deleted_field)}")
+            DELETED_FIELD_FILE.write(bibnb, index, tag, auth_id, deleted_field)
             continue
-        # If no, add the authority ID to the index & the field to the fields
-        auth_id_index.append(auth_id)
-        fields.append(field)
     
-    # Once loop is over, check if there were duplicates for this tag
+    # Once loop is over, for each defined auth_id, add the corretc field
+    for auth_id in auth_id_index:
+        if auth_id_index[auth_id].current_field != None:
+            fields.append(auth_id_index[auth_id].current_field)
+        else:
+            ERRORS_FILE.write(Error_Types.AUTH_ID_HAS_NO_CURRENT_FIELD, index=index, bibnb=bibnb, msg=f"{auth_id} is defined in Index but has no current field")
+            LOG.record_message(Level.ERROR, index, bibnb, msg=f"{auth_id} is defined in Index but has no current field")
+
+    # Once all fields are selected, check if there were duplicates for this tag
     if len(fields) == len(record.get_fields(tag)):
         LOG.record_message(Level.INFO, index, bibnb, msg=f"Tag {tag} did not have duplicates")
         return False
@@ -246,7 +332,7 @@ with open(INPUT_FILE_PATH, mode="r") as f:
             ERRORS_FILE.write(Error_Types.RECORD_WAS_NOT_CHANGED, index=index, bibnb=bibnb)
             LOG.record_message(Level.INFO, index, bibnb, "Record was not changed")
             continue
-
+        
         # If the record was changed, send the edited one to Koha via PUT API
         update_response = KOHA.update_biblio(bibnb, record=record.as_marc())
         # An error occured while getting the record, log & skip to next one